+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@devvit/public-api": "0.11.3",
+    "@devvit/protos": "0.11.3",
     "@devvit/shared-types": "0.11.3",
     "@types/lodash": "^4.14.202",
     "@types/pluralize": "^0.0.33",
